menambahkan fitur login dan register

+++ app/Http/Controllers/Todo/TodoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Todo;
 use App\Http\Controllers\Controller;
 use App\Models\Todo;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TodoController extends Controller
 {
@@ -13,11 +14,12 @@ class TodoController extends Controller
      */
     public function index()
     {
+         
         $max_data = 5;
         if (request('search')) {
-            $data = Todo::where('task','like','%'.request('search').'%')->paginate($max_data)->withQueryString();
+            $data = Todo::where('task','like','%'.request('search').'%')->where('user_id', Auth::user()->id)->orderBy('task', 'asc')->paginate($max_data)->withQueryString();
         }else{
-            $data = Todo::orderBy('task','asc')->paginate($max_data);
+            $data = Todo::orderBy('task','asc')->where('user_id', Auth::user()->id)->paginate($max_data);
         }
         return view('todo.app',compact('data'));
       
@@ -46,6 +48,7 @@ class TodoController extends Controller
 
         $data = [
             'task' => $request->input('task'),
+            'user_id' => Auth::user()->id
         ];
 
         Todo::create($data);
@@ -87,7 +90,7 @@ class TodoController extends Controller
             'is_done'=>$request->input('is_done')
         ];
 
-        Todo::where('id',$id)->update($data);
+        Todo::where('id',$id)->where('user_id', Auth::user()->id)->update($data);
         return redirect()->route('todo')->with('success','berhasil update data');
     }
 
@@ -97,7 +100,7 @@ class TodoController extends Controller
     public function destroy(string $id)
     {
 
-        Todo::where('id',$id)->delete();
+        Todo::where('id',$id)->where('user_id', Auth::user()->id)->delete();
         return redirect()->route('todo')->with('success','berhasil delete data');
     }
 }
